Final Stats Implementation & Pause Menu Fixes
-Data now properly tracks lifetime stats and displays them correctly
-Fixed major issue that would not allow game to be paused when using Bluetooth controller
-Fixed bug where Life Jacket & Rat Poison Souvenirs were mismatched when selected
-Darkened background of HTP Menu

+++ Assets/Scripts/Menu/StateNameController.cs
@@ -32,7 +32,7 @@ public class StateNameController : MonoBehaviour
 
     public static int totalFights;
     public static int totalWins;
-    public static float winRate;
+    public static int winRate;
     public static int[] fightStyleUsage = new int[6];
     //Karate, Boxing, MMA, TKD, Kung Fu, Wrestling
     public static float avgBasicAttackConnectionRate;
